requirements.txt
asgiref==3.2.7
Django==3.0.5
django-admin-tools==0.9.0
django-ckeditor==5.9.0
django-cors-headers==3.2.1
django-js-asset==1.2.2
django-mptt==0.11.0
django-mptt-admin==1.0.1
django-smart-selects==1.5.4
django-treeadmin==0.4.1
html5lib==1.0.1
Pillow==7.1.2
pkg-resources==0.0.0
PyPDF2==1.26.0
pytz==2020.1
reportlab==3.5.42
six==1.14.0
sqlparse==0.3.1
webencodings==0.5.1
xhtml2pdf==0.2.4
